requirements.txt
mock
numpy>=1.6.2
scipy>=0.13.3
liac-arff>=2.1.1dev
xmltodict
nose
requests
scikit-learn